+++ lib/aliases.f
@@ -20,7 +20,7 @@
 
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !!
-!> \file checkin.f
+!> \file aliases.f
 !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -32,37 +32,45 @@
 !> \note        See `LICENSE' for full license.
 !>              See `README.md' for project details.
 !>
-!> \brief   Immediately stage and commit all changes at once.
+!> \brief   The Git alias commands `git-aliases` provides.
 !>
-!> This alias is used in order to stage and commit all changes at once with just
-!> a single Git instruction.  In order to avoid option clashes with other passed
-!> options, the command definition was split into multiple Git operations.
-!>
-!> When working on a project, especially on the command line, it is common to
-!> finalise the changes to commit, at first, and to commit them all at once when
-!> the intended progress was achieved.  Then, this alias command comes in handy
-!> in order to persist them with a single instruction.
-!>
-!> Another possible use case is measuring the runtime of applications.  One can
-!> then use this alias in scripts in order to commit the measured data as soon
-!> as it is available.  That way, a loss of data is excluded due to the design
-!> of Git.
-!>
-!> Due to this command's definition, it is possible to pass additional options
-!> to `git commit` as required.
+!> This is the library submodule containing the Git alias commands `git-aliases`
+!> provides.
 !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-      subroutine checkin
-      use libgaf18, only: git
+      submodule (libgaf18) aliases
       implicit none
+      contains
 
-      call git
-     &( 'checkin'
-     &, '!git preview . && git commit'
-     &, 'Create a new commit immediately.'
-     &)
+      include '../aliases/backup.f'
+      include '../aliases/backupdestination.f'
+      include '../aliases/bone.f'
+      include '../aliases/bugfix.f'
+      include '../aliases/checkin.f'
+      include '../aliases/complain.f'
+      include '../aliases/create.f'
+      include '../aliases/delbranch.f'
+      include '../aliases/ff.f'
+      include '../aliases/goto.f'
+      include '../aliases/grab.f'
+      include '../aliases/graph.f'
+      include '../aliases/newbranch.f'
+      include '../aliases/noff.f'
+      include '../aliases/preview.f'
+      include '../aliases/rewind.f'
+      include '../aliases/savetags.f'
+      include '../aliases/store.f'
+      include '../aliases/subinit.f'
+      include '../aliases/tidy.f'
+      include '../aliases/tws.f'
+      include '../aliases/unconfigure.f'
+      include '../aliases/verbose.f'
+      include '../aliases/view.f'
+      include '../aliases/whereami.f'
+      include '../aliases/whichremotes.f'
+      include '../aliases/withdraw.f'
 
-      end subroutine checkin
+      end submodule aliases
 
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!