+++ CHANGELOG.md
@@ -53,6 +53,7 @@ based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * alias:  `graph`
 * alias:  `newbranch`
 * alias:  `noff`
+* alias:  `preview`
 * alias:  `rewind`
 * alias:  `savetags`
 * alias:  `store`
@@ -74,12 +75,15 @@ based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * create this changelog
 * create this repository
 * define Git invocation subroutine
+* define self information subroutine
 * define summary subroutine for aliases to be configured
+* define table header subroutine
 * Doxygen settings:  Fortran presets
 * GitHub Action:  CFF validation
 * GitHub Issue Template:  new alias
 * license:  GPL-2.0
-* module for alias commands
-* module for business logic
+* module for project logic
+* submodule for alias commands
+* submodule for business logic
 
 <!----------------------------------------------------------------------------->